+++ lib/Crypto/SelfTest/Cipher/__init__.py
@@ -35,7 +35,6 @@ def get_tests(config={}):
     from Crypto.SelfTest.Cipher import test_CAST;     tests += test_CAST.get_tests(config=config)
     from Crypto.SelfTest.Cipher import test_DES3;     tests += test_DES3.get_tests(config=config)
     from Crypto.SelfTest.Cipher import test_DES;      tests += test_DES.get_tests(config=config)
-    from Crypto.SelfTest.Cipher import test_XOR;      tests += test_XOR.get_tests(config=config)
     from Crypto.SelfTest.Cipher import test_Salsa20; tests += test_Salsa20.get_tests(config=config)
     from Crypto.SelfTest.Cipher import test_pkcs1_15; tests += test_pkcs1_15.get_tests(config=config)
     from Crypto.SelfTest.Cipher import test_pkcs1_oaep; tests += test_pkcs1_oaep.get_tests(config=config)